Rename approx_equality.py to approximate_equality.py
